requirements.txt
dbus-python==1.3.2
pulsectl==23.5.2
pycairo==1.25.0
PyGObject==3.46.0
